Add build tags to many commands
Fixes #1464

This also removes the `:GoBuildTags` command. The documentation for
`g:go_build_tags` mentions that "This is a permanent setting. A more
useful way is to use |:GoBuildTags| to dynamically change or remove
build tags", but all `:GoBuildTags` does is change the `go_build_tags`
setting, so I guess that's an artefact from a previous implementation?
We could keep it if people prefer using this, but it seems redundant to
me.

Depends on:

https://github.com/davidrjenni/reftools/pull/14
https://github.com/dominikh/go-tools/pull/272

TODO: Add tests to ensure no silly mistakes were made.